payload/producer/requirements.txt
kafka-python>=2.0.1
msgpack>=1.0.0
tzlocal>=2.1 